requirements.txt
django==4.2
psycopg2==2.9.6
djangorestframework==3.14.0